The four biggest files were a third or more inline tests, so `server/mod.rs`,
`server/rooms.rs`, `net/mod.rs` and `sim/world.rs` now keep their tests in
sibling files the way `sim/rule.rs` always has. The three whole-world passes a
generation runs outside the rule are three files — `world/turrets.rs`,
`world/dynamite.rs`, `world/overclock.rs` — where a pass is a thing you can
read rather than eight hundred lines in the middle of something else. Where a
player starts is `net::spawn`. A match and its sides are `server/matches.rs`.
`answer_for` is one match rather than twelve consecutive if-lets, a room is
asked for with one value, and three long design arguments moved into docs/
leaving the sentence that stops somebody re-introducing the fault.

**Clippy says nothing now, either way it is run.** The last five were mine to
finish: two enums whose largest variant was carried everywhere by everything
that was not it, and three draw functions past the argument cap — which this
tree already had an answer for, since `lobby::Look` exists precisely because
its own list reached eleven. The clock and the stamp panel have a `Look`, and
one square of the hotbar has a `Cap`.

Two things worth saying about the merge itself. `cell.rs` and `rule.rs` are
untouched: the branch forked before the overclocker became a shape that flies,
and both files are exactly as they were. And the split carried the *old*
overclock pass with it, so the Moore-neighbourhood mask and the tests for a
glider at double speed were re-applied into `world/overclock.rs` and
`world/tests.rs` by hand rather than trusting the move.

574 tests, formatting clean, server binaries and the wasm module both build.

+++ src/client/views/menu/attract.rs
@@ -145,7 +145,7 @@ mod tests {
     fn the_soup_is_somewhere_between_empty_and_solid() {
         let it = Attract::new();
         let alive = it.world.live_cells().len();
-        let side = CHUNKS as usize * crate::sim::CHUNK_N as usize;
+        let side = CHUNKS as usize * crate::sim::CHUNK_N;
         let squares = side * side;
         assert!(alive > squares / 8, "the backdrop is nearly empty: {alive} of {squares}");
         assert!(alive < squares / 2, "the backdrop is nearly solid: {alive} of {squares}");
